fix: properly override template storage implementations for remote (#1582)
### Motivation
When api-impl splitting the RemoteTemplateStorage impl was not done
correctly. Methods that handle file streams (in & out) were just called
as usual RPCs which is not possible.

### Modification
This PR overrides the specific method implementations and uses the
chunked file transfer api to transfer filestreams.

### Result
The template storage is working properly from the wrapper side.