fix: provide context class loader when transforming classes (#1601)
### Motivation
When a class is currently being transformed the system class loader is
used to provide access to classes that are required to build the final
class file. However, this causes issues when classes are required that
are not available to the system class loader, e.g. if dynamically loaded
by the server software.

### Modification
Use the class loader of the class that is being transformed, as it must
be able to provide all necessary classes to perform the transformation.

### Result
No more issues during class transforming due to unknown classes.