fix: provide fallback when server dies on bungeecord (#1602)
### Motivation
Currently, when a server unexpectedly closes the connection to the
proxy, the player gets disconnected as CloudNet does not provide a
fallback server. This is due to the fact that the ServerKickEvent is
only called when the server sends a kick packet. In case the server goes
down, BungeeCord tries to connect the player to the first priority (in
our case a dummy `lobby` server as BungeeCord won't start without a
priority) with the connection reason `SERVER_DOWN_REDIRECT`. We can
catch this reason and provide a proper fallback in `ServerConnectEvent`.

### Modification
Catch the `SERVER_DOWN_REDIRECT` reason in `ServerConnectEvent` and
provide a proper fallback. If no fallback is available, disconnect the
player with the new disconnection reason introduced in #1399.

### Result
Players will be provided with a proper fallback in case of an unexpected
connection close by the downstream service instead of being
disconnected.