fix: remove session if chunked file transfer cannot be initiated (#1573)
### Motivation
When a chunked file transfer is started a session is registered into the
local session registry. However, if no data can be provided from the
remote peer (e.g. if a template doesn't exist) the session is not
unregistered from the local session registry.

### Modification
Unregister a session from the local session registry if a chunked file
transfer cannot be started.

### Result
No dead chunked file transfer sessions in local session registry.

---------

Co-authored-by: Pasqual Koschmieder <git@derklaro.dev>