fix: syncproxy module does not reload config file (#1606)
### Motivation
An user reported that the `modules reload CloudNet-SyncProxy` command
would reload the SyncProxy module but the module would not update the
loaded configuration due to a wrongly placed `@Service` annotation.

### Modification
Removed the wrong `@Service` annotation and the unnecessary null check.
The management class is created by aerogel. If aerogel cannot find it /
construct it, aerogel will throw an exception.

### Result
The module configuration is properly reloaded on module reload.

##### Other context
See
https://discord.com/channels/325362837184577536/818777626663321671/1357362934732292187