fix: dockerized-services module image handling (#1621)
### Motivation
The current docker module implementation is not able to handle custom
registries correctly. This is due to 2 issues:
1. The registry was not correctly preprended to image names generated by
`DockerImage#imageName`
2. The PullImageCmd construction used `PullImageCmd#withRegistry` which
doesn't do anything (and never did something, the query option was never
implemented in the docker api as far as I could tell)

### Modification
1. Fix `DockerImage#imageName` to correctly handle registries. I've also
added support for SHA256 image digests to be used instead of image tags.
Also added a test to validate that the outcome of
`DockerImage#imageName` is valid.
2. No longer rely on `PullImageCmd#withRegistry`, supplying the full
docker image name constructed by `DockerImage#imageName` is sufficient
along with supplying the platform.
3. Supply the docker image platform to CreateContainerCmd to actually
select the image for the requested platform.

### Result
Custom docker registries and docker image platforms work as expected.