ci: cleanup/fine-tune the build action (#1630)
### Motivation
The build action needs some fine tuning and cleanup as it does not work
as expected in some scenarios.

### Modification
1. Remove all references to the old github merge queue experiment, this
is no longer needed and is annoying when reading the action.
2. Replace deprecated `junit_files` input of the unit test publish
action and always publish unit tests results. Comments on pull requests
are now disabled unless the action is triggered by a pull request (to
prevent duplicate comments for pushes and the pull request run).
3. The updater metadata is now only published if the action is triggered
by a push or workflow dispatch.
4. Artifacts generated by the action are now always uploaded (there is
no reason to not do it).

### Result
The build action is easier to read and behaves more as expected.