fix: language question result conversion in default setup (#1640)
### Motivation
The code for selecting the language in the default task setup was
changed to use locales instead of strings, however the result handling
was not changed accordingly. This results in a class cast exception when
the results are being processed.

### Modification
Change the default setup result handler to use locales instead of
strings.

### Result
No more class cast exceptions when completing the default config setup.