fix: singleton annotation locations (#1642)
### Motivation
Some singleton annotations were located at the wrong places and should
be moved to their respective factory methods.

### Modification
The singleton annotation for the node configuration was located both on
the implementation and the interface. This is not acceptable as scope
annotations are not allowed on interfaces/abstract classes and the
implementation is not relevant to the binding builder, therefore the
annotation was ignored there anyway. Same applies to the wrapper
configuration.

### Result
The singleton scope annotations are now located at the factory methods
rather than at locations where they don't have any effect.