fix: handle SIGHUP in launcher-patcher to prevent exits before updating (#1641)
### Motivation
The launcher-patcher tries to update the launcher.jar after the node &
launcher exit. This process did not work when `screen` or `tmux` was
used to start the launcher due to the fact that they send a `SIGHUG`
signal and the JVM handles the signal and stops the still running
(waiting) launcher-patcher. This required users to download the launcher
themselves when we did changes to the launcher.

### Modification
Introduced a handler for `SIGHUP` in the launcher-patcher that ignores
the signal. Furthermore, this introduces a 5 second timeout when waiting
for the launcher to exit before running the patcher as the launcher
should not take that long to stop and to prevent replacing the next
running launcher instance getting replaced while running.

### Result
Launcher updates are properly working when using `screen` or `tmux`

---------

Co-authored-by: Pasqual Koschmieder <git@derklaro.dev>