feat: make command confirmations skippable using extra flag (#1645)
### Motivation
In #1579 it was requested to make command confirmations (like needed for
`exit`) skippable using a flag.

### Modification
Added a new `@SkipConfirmation` annotation which allows to specify a
flag name which can be used to skip the confirmation of a command. This
PR also directly includes this annotation for the exit command so that
it can be skipped using `--skip`

### Result
Commands that need to be confirmed can be skipped if requested.

##### Other context
Resolves #1579