feat: allow subnets in node ip allowlist (#1644)
### Motivation
The current node ip allowlist is only able to handle static ip
addresses, not subnets. This causes issues with subnet usage, for
example, when running services in docker containers using a custom
network.

### Modification
Add support for subnets in the node ip allowlist. The allowlist accepts
static ips or subnets, depending on the fact if the allowlist entry
contains a `/` to delimit the ip from the subnet prefix length.

### Result
The ip allowlist can now handle subnets alongside static ip addresses.