fix: make older asm versions support java 24 (#1653)
### Motivation
Older paper & spigot versions bundle asm versions that do not support
java 24. Paper uses asm to remap plugins which are spigot mapped to
mojang mappings. This fails due to the asm version mismatch. This PR
fixes that.

### Modification
Added a class transformer for the asm ClassReader which removes the Java
version check. The check is not needed as ASM is able to still rewrite
newer classes.

### Result
Remapping on older paper versions with older asm versions works on Java
24

---------

Co-authored-by: Pasqual Koschmieder <git@derklaro.dev>