improve: wrapped application invocation exception handling (#1639)
### Motivation
When starting e.g. a BungeeCord proxy and the proxy fails to start due
to missing native libraries. Currently CloudNet would not catch that the
service is actually stopped and would display it as "RUNNING".

### Modification
Handle exceptions thrown after invoking the main method and halt the
wrapper jvm.

### Result
No services are displayed as running after they crashed in the start
period.

---------

Co-authored-by: Pasqual Koschmieder <git@derklaro.dev>