feat: allow node host override in docker module (#1659)
### Motivation
The ip/port that can be used from inside a docker network to connect to
a node listener on the host system might be different that any
configured listener on the node (e.g. `host.docker.internal` in some
cases). There is currently no way to connect to a node listener outside
the docker network (except for custom modules/server implementations).

### Modification
Add the possibility to override the node host (globally and
task-specific) to allow specifying a different host and port to connect
to from inside a docker network.

### Result
Services can connect to a node listener bound outside of a docker
network.