feat: custom netty memory-segment based buffer impl (#1655)
### Motivation
We're currently using the default memory segment backed buffer
implementation provided by netty. While this implementation works, there
are still some issues with it. These were resolved in upstream but never
published due to the central-portal migration that wasn't done for
netty5 yet.

### Modification
Add a custom netty buffer implementation that is backed by memory
segments with a few modifications:
1. memory segments are allocated manually using `malloc` and freed using
`free`. This means that we're not using the Arena implementations
provided by the jdk (specifically the shared arena), as closing them
provides a performance bottleneck due to a required thread handshake to
prevent uaf. This can be quite expensive depending on the count of
active threads.
2. the bugfixes of the upstream implementation are now already present
in our custom implementation. This means that the buffer should now work
bug-free.
3. the implementation contains some other cleanups that are now possible
after the jdk feature became permanent, e.g. preventing segment slicing
when copying segments.

### Result
We're using a custom mememory-segment backed buffer implementation that
doesn't contain some upstream bugs and is faster to deallocate.