fix: minestom permission checking (#1658)
### Motivation
During the api-impl split we introduced a MinestomPermissionChecker
service as Minestom does not have any permission checking capabilities
anymore. The service discovery does not work properly as the service is
resolved before the service was registered.

### Modification
Use a provider to allow lazy construction of the PermissionChecker.

### Result
The MinestomPermissionChecker is properly resolved when needed.