feat: use java 17 for api components (#1650)
### Motivation
We want to ease the development of modules and plugins that interface
with CloudNet. Currently CloudNet requires Java 24 in its runtime and
consequently its API.
The API has already been extracted from the implementation and thus
efforts to relax the Java version constraints can be realised.


### Modification
With this PR the API components relax their Java version requirements
from Java 24 to Java 17.

### Result
The API components can now be used in projects that do not work or want
to constraint themselves to Java 24.

##### Other context
See this very small exchange on Discord:
https://discord.com/channels/325362837184577536/818777626663321671/1380941029741694986.
These changes built upon #1566.

---------

Co-authored-by: 0utplay <aldin@sijamhodzic.de>
Co-authored-by: Pasqual Koschmieder <git@derklaro.dev>
Co-authored-by: Pasqual Koschmieder <pasqual.koschmieder@easybill.de>
Co-authored-by: Pasqual Koschmieder <pasqual.koschmieder@gmail.com>