feat: enhance sign module compat with new minecraft versions (#1665)
### Motivation
New minecraft versions introduced new sign types and options that aren't
yet supported in the sign module. These are:
1. hanging signs (facing wasn't determined correctly yet for all sign
types)
2, sign text from both sides of a sign (relevant for hanging signs)

### Modification
1. Replace usage of reflexion with bare method handles. This allows for
better performance which is relevant as all methods in
`BukkitCompatibility` are called on the main server thread.
2. Support retrieval of sign faces for all types of signs.
3. Support the new `SignSide` bukkit api and apply text (and other
options) to both sides of a sign instead of just the front.

### Result
Sign changes in recent minecraft/bukkit versions are now supported by
the signs module.