feat: allow modifications to log levels using properties (#1656)
### Motivation
CloudNet provides a logback configuration which works for most users.
But if someone wants to change the log level of a single logger the user
would either need to change the level for all loggers or override the
whole configuration.

### Modification
Added support for logbacks
https://logback.qos.ch/manual/configuration.html#propertiesConfigurator
which allows to set log levels using a simple properties file.

### Result
Log levels can easily be changed by the user without replacing (and
potentially breaking) the whole logback configuration.