chore: improve sign layout ticking (#1668)
### Motivation
The current sign layout ticking process is based on the assumption that
there won't be any race conditions. However, there is a slim chance that
a race happens which will then cause all layouts of a holder to freeze
until a server restart or config reload.

### Modification
Base the sign ticking process and layout retrieval entirely on the last
tick when the sign was updated. This removes the need for different
places to lock/unlock sign layout ticking and therefore prevents races
between the sign ticking thread and the main server thread. This also
applies to the map that tracks the layouts that need to be updated as a
result of the current tick.

### Result
Sign layout ticking is resistent against race conditions and uses the
server tick count as the single point of truth when retrieving the
applicable sign layout.