feat: conditionally disable netty unsafe to leverage the ffm api (#1664)
### Motivation
In netty 4.2.3 a new type of ByteBuffer cleaner (or allocator) will be
introduced that uses `malloc` and `free` directly using the ffm api on
java 24+. In order to enable this cleaner, unsafe has to be
unavailable/disabled. Since this can be a performance issue (as the
ByteBuffer lifecycle might be controlled by the garbage collector if
unsafe is unavailable) we only disable unsafe when the new cleaner
implementation is available.

### Modification
Add a transformer that checks if all conditions are met to disable
unsafe in order to enable the ffm-based cleaner implementation. If that
is the case, the `PlatformDependent0#explicitNoUnsafeCause0()` method
will be overridden to always return an UOE to indicate that unsafe is
explicitly disabled.

### Result
On netty 4.2.3+ unsafe will be disabled in favor of the ffm-linker based
cleaner implementation, but only is all necessary preconditions to
enable the cleaner are met.