fix: only handle npc interact when performed with main hand (#1673)
### Motivation
The client sends the interact packet twice (in most cases): once for the
main hand and once for the off hand. This causes all right click actions
of selector entities to be executed twice.

### Modification
Only actually handle the right click action to a selector entity if the
interact happened with the main hand. In all other cases we just ignore
the interact and, if needed, just cancel the interact event (required to
cancel certain actions on mobs such as shearing a sheep).

### Result
Selector entity right click actions are only performed once, not twice
in many cases.

##### Other context

https://canary.discord.com/channels/325362837184577536/818777626663321671/1388872733903028265