chore: migrate to paper downloads api v3 (#1671)
### Motivation
Paper recently released a new version of their downloads api and we
should migrate to it.

### Modification
Migrate all uses of the current v2 api to the new v3 api. This also uses
the newly introduced api endpoint to get the latest version directly,
instead of sorting the available build numbers. The download url is now
extracted from the api response as well.

### Result
The new v3 paper downloads api version is used instead of the v2
version.