fix: unsafe transformer disable system property (#1681)
### Motivation
The unsafe transformer is registered in a different way than usual
(usually transformers are registered via SPI), so disabling the
transformer by throwing an exception is not a viable option as it
results in a jvm crash.

### Modification
Move the transformer disabled check into the premain class and don't try
to construct/register the transformer if it is disabled.

### Result
No more jvm crashes when the unsafe transformer is disabled using a
system property.