feat: split sign text color and glowing setting (#1670)
### Motivation
Since minecraft 1.14 the text color of a sign can be changed and since
1.17 the text of a sign can glow. We currently only support both options
in one setting, which should be split up into two.

### Modification
Add a setting to set the text color of a sign and one to set the glowing
state of a sign. Also fixes the sign handling on minestom, as it was
still using the pre-1.20 sign nbt format.

### Result
The text color of a sign can be configured separately from the fact if
the sign text is glowing.