fix: compute field offsets per-class instead of globally (#1683)
### Motivation
Currently our unsafe replacement implementation computes field offsets
using a global long counter (which should be fine as the documentation
in `sun.misc.Unsafe` is not specifying how field offsets are resolved).
However, there are cases where libraries depend on the fact that the
returned offset is rather small and computed based on the class tree.
This is for example the case in the google protobuf implementation,
which allow a max field offset value of `(1 << 20) - 1` (1,048.575) as
they do some optimizations using bit packing.

### Modification
Compute field offsets based on the class hierachy instead of globally
and provide values that are much closer to the values that are actually
returned by `sun.misc.Unsafe`.

### Result
No more breaking library code because of bad assumptions about the
return value of field offset getter methods.

##### Other context
Code that packs the field offset and breaks:
https://github.com/protocolbuffers/protobuf/blob/960e79087b332583c80537c949621108a85aa442/java/core/src/main/java/com/google/protobuf/MessageSchema.java#L568-L575