chore: fetch all paper downloads from api (#1684)
### Motivation
Currently the download urls of old paper versions are hardcoded in the
`versions.json` file, which shouldn't be done. As the paper api v3
update announcement states: `[...] and we advise against doing so, as
the format may change.`

### Modification
Always request the download url for a version from the paper downloads
api instead of hardcoding them.

### Result
Download urls of old paper versions are resolved from the paper api
instead of being hardcoded, as recommended in the update announcement.