fix: override pull(Async) in RemoteTemplateStorage (#1692)
### Motivation
RemoteTemplateStorage does currently not override `pull` and `pullAsync`
leading to the methods being executed on the node. This leads to an
issue where the template is not pulled into the a relative path to the
current wrapper directory, but relative to the node directory.

### Modification
Override `pull` and `pullAsync` to let them use the chunked data
transfer infrastructure and extract the zipped template into the
(possible relative to the service directory) path passed to the
respective `pull` method.

### Result
RemoteTemplateStorage has a correct implementation of `pull` and
`pullAsync`.

##### Other context
Fixes #1303