chore: allow re-using channel message builders again (#1695)
### Motivation
Currently, a builder for a ChannelMessage cannot be re-used. This is
done because the content buffer provided to the builder shouldn't be
shared across multiple channel message instances at the same time.
However, this limits the ability to re-use a base builder that just has
changing content.

### Modification
Instead of holding a state if the `build` was called before, the build
methods now take in the content buffer for the channel message. While
this is not a perfect prevention against a buffer re-use, it at least
limits the possibility of doing it by accident while also allowing the
base of a channel message build to be reused.

### Result
The chance of reusing the content buffer of a channel message on
accident has been reduced, while also allowing to reuse a base channel
message builder instance.