chore: reduce heap usage when doing chunked data transfer (#1702)
### Motivation
Currently, chunked data transfers are copying the data via the heap into
the target netty buffer (and vice-versa). This is very inefficient as
this means that a native memory segment is copied to the heap and then
back into a native memory segment.

### Modification
Use nio Channels in combination with ByteBuffers to directly read/write
the content being transferred to/from the netty buffer, completely
removing the need for (almost) any heap allocations.

### Result
Reduced heap usage when doing a chunked data transfers.