chore: prevent invalid packets/network state (#1703)
### Motivation
It is currently possible that an invalid network state is caused by an
invalid packet length being sent from one of the connected clients.
Furthermore, an exception can be triggered when trying to connect to a
node listener directly through a minecraft client, which isn't very
helpful.

### Modification
When an invalid packet length is received the connection to the sender
is closed immediately. Furthermore, the packet decoder implementation no
longer starts cumulating packet buffers when there is a decoding
failure.
Each packet sent by CloudNet is now prefixed with a magic integer. If an
incoming packet is not prefixed with this magic number, the connection
to the client is closed. This prevents connections that are sending
packets not formatted in the CloudNet protocol format (e.g., the
minecraft client) to cause decoding exceptions.

### Result
Better protection against unexpected internal faults and better
information for user faults.