chore: log file query transfer exceptions (#1704)
### Motivation
When sending a file as a response to a query (e.g., from a
RemoteTemplateStorage) any thrown error is just silently discarded and a
memory leak is introduced as the chunked transfer info is not closed.

### Modification
Log any errors thrown during the response transfer and properly close
the chunked transfer session info.

### Result
No more silently ignored errors and another possible memory leak
resolved.