feat: support all fabric server versions (#1719)
### Motivation
There were a lot of questions asking about support for more than the
latest fabric version.

### Modification
Use the nested mods feature of the fabric loader to load
version-specific mods (nested mods are only loaded if possible depending
on their version constraints). In all cases where no version mod is
available, a fallback listener is registered on the services to mark
them as online after a small delay to allow for the server to start.
This comes with the downside that no player count and player data can be
provided to the api, but that should not be much of a problem.

### Result
All versions of minecraft (using the fabric mod loader) are now
supported by the bridge module.