refactor: optimize field operations in unsafe replacement (#1733)
### Motivation
Currently the replacement methods for field access are implemented very
inefficient, leading to performance problems when a lot of calls are
being made to them.

### Modification
1. Migrated away from passing around fields to wrapping them into a
FieldAccessor instance which caches a VarHandle and optionally a
MethodHandle instance for accessing/modifying the wrapped field.
2. Changed the FieldOffsetOps to cache these FieldAccessors instead of
caching the Fields.
3. Changed the FieldOffsetOps to use a lock-free caching solution based
on ClassValue and a FieldAccessor array, rather than using a
ConcurrentHashMap with a custom key type.
4. Reduced the complexity when using the VarHandle/MethodHandle instance
for a field by introducing a dummy Object argument for static fields to
allow passing an instance when none is actually needed.

### Result
The unsafe replacement methods for field access improved significantly
leading to way better server performance under certain circumstances.