fix: respect array offsets in MemorySegmentBuffer (#1737)
### Motivation
Currently the backing array offset of a MemorySegment is not respected
when reading bytes from a MemorySegmentBuffer. This results in random
bytes being read instead of the bytes actually associated with the
buffer. This issue was previously not detected in tests as the offset
only becomes >0 when the backing MemorySegment is sliced, pooled or
based off another java.nio.Buffer.

### Modification
Respect the offset into the array wrapped in a MemorySegment to read the
bytes associated with the current buffer. Also adds a tests that slices
the wrapped MemorySegment making the offset into the array >0.

### Result
No more read of bytes from the backing MemorySegment array outside of
the current buffer.

##### Other context
Fixes CloudNetService/module-rest#123