[ci skip] docs: push repo readme as dockerhub description (#1775)
### Motivation
The current readme should link to our published image on dockerhub and
should be included as the description of the image on dockerhub as well.

### Modification
Add Dockerhub url to readme, push readme as Dockerhub description.

### Result
The readme contains a link to our Dockerhub and the readme is published
as the Dockerhub description when a new image is pushed.