fix: depend on official cloud dependencies in node-api (#1780)
### Motivation
The node-api currently depends on our custom cloud fork which is not
ideal as it requires an additional repository to be defined when using
the dependency with gradle.

### Modification
Use the official cloud-api in the node-api module (changes in our fork
are only implementation details).

### Result
The node-api can be used without specifying additional repositories.