chore: bump gulf to 0.0.1, add cloud-confirmation to node-api (#1781)
### Motivation
We need to use stable dependency versions when publishing to maven
central, therefore we must update to gulf 0.0.1. Also, consumers of the
node-api should be able to use cloud-confirmation, therefore we should
include it in the cloud-api bundle.

### Modification
Bump gulf to 0.0.1, include cloud-confirmation in cloud-api bundle.

### Result
We're using a stable version of gulf and cloud-confirmation is available
for consumers of the node-api package.