fix: add server name to lp context if not otherwise configured (#1802)
### Motivation
The server key is an essential part of the luckperms context
calculation. By default, it is taken from the LuckPerms config file,
however, if the default value `global` is left unchanged, the key isn't
added to the context. However, we do know the current server name and
are able to add it to the LuckPerms context.

### Modification
Add the server key holding the current service name to the luckperms
context if the server name is not specified in the LuckPerms
configuration file.

### Result
The server context key is correctly configured automatically by default.
If the user wants to he can still change the value in the LuckPerms
configuration to his needs.