feat: move smart configs into their own files (#1808)
### Motivation
We want to redo the service / task property handling at some point. This
is a pain in the ass if the smart config stays inside the task
properties. Furthermore it simply does not make sense to keep it there
anymore as smart is mostly used on single tasks and not on all of them,
just creating clutter and overhead.

### Modification
Moved each smart config into a separate file in the
modules/CloudNet-Smart directory, if the config was enabled. Disabled
configs are **not migrated**. Old configs are removed from the task
props.

### Result
Smart configs are not part of the task props anymore and the property
rework can be done.