fix: avoid OOM caused by NULL pointer returned from malloc(0) (#1822)
### Motivation
It's left to the implementations of malloc if a `malloc(0)` returns a
NULL pointer or a non-null pointer. However, our validation logic checks
if a NULL pointer is returned and treats it as an OOM.

### Modification
Always pass at least 1 to malloc, ensuring that we only get a NULL
pointer if we're really out of memory.

### Result
No more OOM errors thrown because of `malloc(0)` calls returning a NULL
pointer.